refactor: use-custom-overlay DX 개선
