+++ hooks/use-custom-overlay/types/index.ts
@@ -1,4 +1,3 @@
 export type CreateOverlayElement = (props: {
-  isOpen: boolean;
   close: () => void;
 }) => JSX.Element;
